requirements.txt
# LLM API clients
openai>=1.6.0
anthropic>=0.8.0
google-generativeai>=0.3.0
deepseek-ai>=0.0.2

# Core dependencies
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.3.5
pydantic>=2.4.2
httpx>=0.24.1
tiktoken>=0.4.0
tqdm>=4.66.1

# Utility packages
numpy>=1.20.0
regex>=2023.10.3

# System packages
psutil>=5.9.0

# Environment variables handling
python-dotenv>=1.0.0

# Typing and data handling
pydantic>=2.0.0
typing-extensions>=4.8.0 